Refactor cli.py (5,267 lines) into focused package modules
Split the monolithic insideLLMs/cli.py into a well-organized package:

- cli/_output.py: Colors, colorize, print_*, ProgressBar, Spinner
- cli/_record_utils.py: JSONL I/O, record keys/labels, fingerprinting
- cli/_report_builder.py: ExperimentResult reconstruction from records
- cli/_parsing.py: create_parser() and argument definitions
- cli/commands/*.py: one module per CLI subcommand (15 commands)
- cli/__init__.py: re-exports for full backward compatibility
- cli/__main__.py: supports python -m insideLLMs.cli

All 5,049 tests pass with no import changes needed in test files.
