+++ insideLLMs/cli/__main__.py
@@ -1,0 +1,7 @@
+"""Allow ``python -m insideLLMs.cli`` to work."""
+
+import sys
+
+from insideLLMs.cli import main
+
+sys.exit(main())
